+++ torchmetrics/functional/text/rouge.py
@@ -12,64 +12,150 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import re
-from typing import Dict, List, Tuple, Union
+from collections import defaultdict
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 import torch
-from torch import Tensor, tensor
+from torch import Tensor
 
-from torchmetrics.utilities.imports import _NLTK_AVAILABLE, _ROUGE_SCORE_AVAILABLE
+from torchmetrics.utilities.imports import _NLTK_AVAILABLE
 
-if _ROUGE_SCORE_AVAILABLE:
-    from rouge_score.rouge_scorer import RougeScorer
-    from rouge_score.scoring import AggregateScore, BootstrapAggregator
-else:
-    RougeScorer, AggregateScore, BootstrapAggregator = object, object, object
-
-ALLOWED_ROUGE_KEYS = (
-    "rouge1",
-    "rouge2",
-    "rouge3",
-    "rouge4",
-    "rouge5",
-    "rouge6",
-    "rouge7",
-    "rouge8",
-    "rouge9",
-    "rougeL",
-    "rougeLsum",
-)
+ALLOWED_ROUGE_KEYS: Dict[str, Union[int, str]] = {
+    "rouge1": 1,
+    "rouge2": 2,
+    "rouge3": 3,
+    "rouge4": 4,
+    "rouge5": 5,
+    "rouge6": 6,
+    "rouge7": 7,
+    "rouge8": 8,
+    "rouge9": 9,
+    "rougeL": "L",
+    "rougeLsum": "Lsum",
+}
 
 
-def add_newline_to_end_of_each_sentence(x: str) -> str:
+def _add_newline_to_end_of_each_sentence(x: str) -> str:
     """This was added to get rougeLsum scores matching published rougeL scores for BART and PEGASUS."""
-    if _NLTK_AVAILABLE:
-        import nltk
+    if not _NLTK_AVAILABLE:
+        raise ValueError("ROUGE-Lsum calculation requires that nltk is installed. Use `pip install nltk`.")
+    import nltk
 
-        nltk.download("punkt", quiet=True, force=False)
+    nltk.download("punkt", quiet=True, force=False)
 
     re.sub("<n>", "", x)  # remove pegasus newline char
-    assert nltk, "nltk must be installed to separate newlines between sentences. (pip install nltk)"
     return "\n".join(nltk.sent_tokenize(x))
 
 
-def format_rouge_results(result: Dict[str, AggregateScore], decimal_places: int = 4) -> Dict[str, Tensor]:
-    """Formats the computed (aggregated) rouge score to a dictionary of tensors format."""
-    flattened_result = {}
-    for rouge_key, rouge_aggregate_score in result.items():
-        for stat in ["precision", "recall", "fmeasure"]:
-            mid = rouge_aggregate_score.mid
-            score = round(getattr(mid, stat), decimal_places)
-            flattened_result[f"{rouge_key}_{stat}"] = tensor(score, dtype=torch.float)
-    return flattened_result
+def _compute_metrics(hits_or_lcs: int, pred_len: int, target_len: int) -> Dict[str, float]:
+    """This computes precision, recall and F1 score based on hits/lcs, and the length of lists of tokenizer
+    predicted and target sentences.
+
+    Args:
+        hits_or_lcs:
+            A number of matches or a length of the longest common subsequence.
+        pred_len:
+            A length of a tokenized predicted sentence.
+        target_len:
+            A length of a tokenized target sentence.
+    """
+    precision = hits_or_lcs / pred_len
+    recall = hits_or_lcs / target_len
+    if precision == recall == 0.0:
+        return dict(precision=0.0, recall=0.0, fmeasure=0.0)
+
+    fmeasure = 2 * precision * recall / (precision + recall)
+    return dict(precision=precision, recall=recall, fmeasure=fmeasure)
+
+
+def _lcs(pred_tokens: List[str], target_tokens: List[str]) -> int:
+    """Common DP algorithm to compute the length of the longest common subsequence.
+
+    Args:
+        pred_tokens:
+            A tokenized predicted sentence.
+        target_toknes:
+            A tokenized target sentence.
+    """
+    LCS = [[0] * (len(pred_tokens) + 1) for _ in range(len(target_tokens) + 1)]
+    for i in range(1, len(target_tokens) + 1):
+        for j in range(1, len(pred_tokens) + 1):
+            if target_tokens[i - 1] == pred_tokens[j - 1]:
+                LCS[i][j] = LCS[i - 1][j - 1] + 1
+            else:
+                LCS[i][j] = max(LCS[i - 1][j], LCS[i][j - 1])
+    return LCS[-1][-1]
+
+
+def _normalize_text(text: str, stemmer: Optional[Any] = None) -> str:
+    """Rouge score should be calculated only over lowercased words and digits. Optionally, Porter stemmer can be
+    used to strip word suffixes to improve matching. The text normalization follows the implemantion from
+    https://github.com/google-research/google-research/blob/master/rouge/tokenize.py.
+
+    Args:
+        text:
+            An input sentence.
+        stemmer:
+            Porter stemmer instance to strip word suffixes to improve matching.
+    """
+    text = re.sub(r"[^a-z0-9]+", " ", text.lower())
+    if stemmer:
+        text = " ".join(stemmer.stem(x) if len(x) > 3 else x for x in text.split())
+    return text.strip()  # to ensure there are no whitespaces as the end of sentence
+
+
+def _rouge_n_score(pred: str, target: str, n_gram: int) -> Dict[str, float]:
+    """This computes precision, recall and F1 score for the Rouge-N metric.
+
+    Args:
+        pred:
+            A predicted sentence.
+        target:
+            A target sentence.
+        n_gram:
+            N-gram overlap.
+    """
+    pred_tokenized, target_tokenized = _tokenize(pred, n_gram), _tokenize(target, n_gram)
+    pred_len, target_len = len(pred_tokenized), len(target_tokenized)
+    if 0 in (pred_len, target_len):
+        return dict(precision=0.0, recall=0.0, fmeasure=0.0)
+
+    pred_counter: Dict[str, int] = defaultdict(int)
+    target_counter: Dict[str, int] = defaultdict(int)
+    for w in pred_tokenized:
+        pred_counter[w] += 1
+    for w in target_tokenized:
+        target_counter[w] += 1
+    # It is sufficient to take a set(pred_tokenized) for hits count as we consider intersenction of pred & target
+    hits = sum(min(pred_counter[w], target_counter[w]) for w in set(pred_tokenized))
+    return _compute_metrics(hits, pred_len, target_len)
+
+
+def _rouge_l_score(pred: str, target: str) -> Dict[str, float]:
+    """This computes precision, recall and F1 score for the Rouge-L or Rouge-LSum metric.
+
+    Args:
+        pred:
+            A predicted sentence.
+        target:
+            A target sentence.
+    """
+    pred_tokenized, target_tokenized = _tokenize(pred, 1), _tokenize(target, 1)
+    pred_len, target_len = len(pred_tokenized), len(target_tokenized)
+    if 0 in (pred_len, target_len):
+        return dict(precision=0.0, recall=0.0, fmeasure=0.0)
+
+    lcs = _lcs(pred_tokenized, target_tokenized)
+    return _compute_metrics(lcs, pred_len, target_len)
 
 
 def _rouge_score_update(
     preds: List[str],
     targets: List[str],
-    scorer: RougeScorer,
-    aggregator: BootstrapAggregator,
-    newline_sep: bool = False,
-) -> None:
+    rouge_keys_values: List[Union[int, str]],
+    results: Optional[Dict[Union[int, str], List[Dict[str, float]]]] = None,
+    stemmer: Optional[Any] = None,
+) -> Dict[Union[int, str], List[Dict[str, float]]]:
     """Update the rouge score with the current set of predicted and target sentences.
 
     Args:
@@ -77,49 +163,69 @@ def _rouge_score_update(
             An iterable of predicted sentences.
         targets:
             An iterable of target sentences.
-        scorer:
-            An instance of the ``RougeScorer`` class from the ``rouge_score`` package.
-        aggregator:
-            An instance of the ``BootstrapAggregator`` from the from the ``rouge_score`` package.
-        newline_sep:
-            New line separate the inputs.
+        rouge_keys_values:
+            List of N-grams/'L'/'Lsum' arguments.
+        stemmer:
+            Porter stemmer instance to strip word suffixes to improve matching.
 
     Example:
         >>> targets = "Is your name John".split()
         >>> preds = "My name is John".split()
-        >>> aggregator = BootstrapAggregator()
-        >>> scorer = RougeScorer(rouge_types=("rouge1", "rouge2", "rougeL", "rougeLsum"), use_stemmer=False)
-        >>> _rouge_score_update(preds, targets, scorer=scorer, aggregator=aggregator, newline_sep=False)
+        >>> from pprint import pprint
+        >>> score = _rouge_score_update(preds, targets, rouge_keys_values=[1, 2, 3, 'L'])
+        >>> pprint(score)  # doctest: +NORMALIZE_WHITESPACE +SKIP
+        {'1': {'precision': 0.25, 'recall': 0.25, 'fmeasure': 0.25},
+         '2': {'precision': 0.0, 'recall': 0.0, 'fmeasure': 0.0},
+         '3': {'precision': 0.0, 'recall': 0.0, 'fmeasure': 0.0},
+         'L': {'precision': 0.25, 'recall': 0.25, 'fmeasure': 0.25}}
     """
-    for pred, target in zip(preds, targets):
+    results = results if results is not None else {rouge_key: [] for rouge_key in rouge_keys_values}
+    for pred_raw, target_raw in zip(preds, targets):
+        pred, target = _normalize_text(pred_raw, stemmer), _normalize_text(target_raw, stemmer)
         # rougeLsum expects "\n" separated sentences within a summary
-        if newline_sep:
-            pred = add_newline_to_end_of_each_sentence(pred)
-            target = add_newline_to_end_of_each_sentence(target)
-        results = scorer.score(pred, target)
-        aggregator.add_scores(results)
+        if "Lsum" in rouge_keys_values:
+            pred_sum = _normalize_text(_add_newline_to_end_of_each_sentence(pred_raw), stemmer)
+            target_sum = _normalize_text(_add_newline_to_end_of_each_sentence(target_raw), stemmer)
+
+        for rouge_key in rouge_keys_values:
+            if isinstance(rouge_key, int):
+                score = _rouge_n_score(pred, target, rouge_key)
+            else:
+                score = _rouge_l_score(
+                    pred if rouge_key != "Lsum" else pred_sum,
+                    target if rouge_key != "Lsum" else target_sum,
+                )
+            results[rouge_key].append(score)
+    return results
 
 
-def _rouge_score_compute(aggregator: BootstrapAggregator, decimal_places: int = 4) -> Dict[str, Tensor]:
+def _rouge_score_compute(
+    sentence_results: Optional[Dict[Union[int, str], List[Dict[str, float]]]]
+) -> Dict[str, Tensor]:
     """Compute the combined ROUGE metric for all the input set of predicted and target sentences.
 
     Args:
-        aggregator:
-            An instance of the ``BootstrapAggregator`` from the from the ``rouge_score`` package.
-        decimal_places:
-            The number of digits to round the computed the values to.
+        sentence_results:
+            Rouge-N/Rouge-L/Rouge-LSum metrics calculated for single sentence.
     """
-    result = aggregator.aggregate()
-    return format_rouge_results(result, decimal_places)
+    results: Dict[str, Tensor] = {}
+    # Obtain mean scores for individual rouge metrics
+    if sentence_results is None:
+        return results
+    for rouge_key, scores in sentence_results.items():
+        res = torch.tensor([(score["precision"], score["recall"], score["fmeasure"]) for score in scores]).mean(0)
+        results[f"rouge{rouge_key}_precision"] = res[0]
+        results[f"rouge{rouge_key}_recall"] = res[1]
+        results[f"rouge{rouge_key}_fmeasure"] = res[2]
+
+    return results
 
 
 def rouge_score(
     preds: Union[str, List[str]],
     targets: Union[str, List[str]],
-    newline_sep: bool = False,
     use_stemmer: bool = False,
     rouge_keys: Union[str, Tuple[str, ...]] = ("rouge1", "rouge2", "rougeL", "rougeLsum"),  # type: ignore
-    decimal_places: int = 4,
 ) -> Dict[str, Tensor]:
     """Calculate `ROUGE score <https://en.wikipedia.org/wiki/ROUGE_(metric)>`_, used for automatic summarization.
 
@@ -128,15 +234,11 @@ def rouge_score(
             An iterable of predicted sentences.
         targets:
             An iterable of target sentences.
-        newline_sep:
-            New line separate the inputs.
         use_stemmer:
             Use Porter stemmer to strip word suffixes to improve matching.
         rouge_keys:
             A list of rouge types to calculate.
             Keys that are allowed are ``rougeL``, ``rougeLsum``, and ``rouge1`` through ``rouge9``.
-        decimal_places:
-            The number of digits to round the computed the values to.
 
     Return:
         Python dictionary of rouge scores for each input rouge key.
@@ -161,7 +263,7 @@ def rouge_score(
 
     Raises:
         ValueError:
-            If the python packages ``nltk`` or ``rouge-score`` are not installed.
+            If the python package ``nltk`` is not installed.
         ValueError:
             If any of the ``rouge_keys`` does not belong to the allowed set of keys.
 
@@ -169,17 +271,19 @@ def rouge_score(
         [1] ROUGE: A Package for Automatic Evaluation of Summaries by Chin-Yew Lin https://aclanthology.org/W04-1013/
     """
 
-    if not (_NLTK_AVAILABLE and _ROUGE_SCORE_AVAILABLE):
-        raise ValueError(
-            "ROUGE metric requires that both nltk and rouge-score is installed."
-            " Either as `pip install torchmetrics[text]` or `pip install nltk rouge-score`"
-        )
+    if use_stemmer:
+        if not _NLTK_AVAILABLE:
+            raise ValueError("Stemmer requires that nltk is installed. Use `pip install nltk`.")
+        import nltk
+
+    stemmer = nltk.stem.porter.PorterStemmer() if use_stemmer else None
 
     if not isinstance(rouge_keys, tuple):
         rouge_keys = tuple([rouge_keys])
     for key in rouge_keys:
-        if key not in ALLOWED_ROUGE_KEYS:
-            raise ValueError(f"Got unknown rouge key {key}. Expected to be one of {ALLOWED_ROUGE_KEYS}")
+        if key not in ALLOWED_ROUGE_KEYS.keys():
+            raise ValueError(f"Got unknown rouge key {key}. Expected to be one of {list(ALLOWED_ROUGE_KEYS.keys())}")
+    rouge_keys_values = [ALLOWED_ROUGE_KEYS[key] for key in rouge_keys]
 
     if isinstance(preds, str):
         preds = [preds]
@@ -187,8 +291,19 @@ def rouge_score(
     if isinstance(targets, str):
         targets = [targets]
 
-    aggregator = BootstrapAggregator()
-    scorer = RougeScorer(rouge_keys, use_stemmer=use_stemmer)
+    sentence_results = _rouge_score_update(preds, targets, rouge_keys_values, stemmer=stemmer)
+    return _rouge_score_compute(sentence_results)
 
-    _rouge_score_update(preds, targets, scorer=scorer, aggregator=aggregator, newline_sep=newline_sep)
-    return _rouge_score_compute(aggregator=aggregator, decimal_places=decimal_places)
+
+def _tokenize(text: str, n_gram: int) -> List[str]:
+    """Retun the list of a tokenized input text, where tokens are represented by N-grams.
+
+    Args:
+        text:
+            An input sentence.
+        n_gram
+            N-gram size to return.
+    """
+    tokens = re.split(r"\s+", text)
+    n_grams_list = [" ".join(tokens[i : i + n_gram]) for i in range(len(tokens) - n_gram + 1)]
+    return n_grams_list